+++ lib/ts/recipe/emailpassword/recipe.ts
@@ -40,6 +40,8 @@ import { BaseRequest, BaseResponse } from "../../framework";
 import OverrideableBuilder from "supertokens-js-override";
 import EmailDeliveryIngredient from "../../ingredients/emaildelivery";
 import { TypeEmailPasswordEmailDeliveryInput } from "./types";
+import SessionError from "../session/error";
+import { logDebugMessage } from "../../logger";
 
 export default class Recipe extends RecipeModule {
     private static instance: Recipe | undefined = undefined;
@@ -244,7 +246,11 @@ export default class Recipe extends RecipeModule {
     getEmailForUserId = async (userId: string, userContext: any) => {
         let userInfo = await this.recipeInterfaceImpl.getUserById({ userId, userContext });
         if (userInfo === undefined) {
-            throw Error("Unknown User ID provided");
+            logDebugMessage("getEmailForUserId: returning UNAUTHORISED because userInfo is undefined");
+            throw new SessionError({
+                type: SessionError.UNAUTHORISED,
+                message: "Unknown User ID provided",
+            });
         }
         return userInfo.email;
     };