+++ lib/ts/recipe/thirdparty/utils.ts
@@ -19,6 +19,8 @@ import { TypeInput as TypeNormalisedInputEmailVerification } from "../emailverif
 import { RecipeInterface, APIInterface, TypeProvider } from "./types";
 import { TypeInput, TypeNormalisedInput, TypeInputSignInAndUp, TypeNormalisedInputSignInAndUp } from "./types";
 import BackwardCompatibilityService from "./emaildelivery/services/backwardCompatibility";
+import SessionError from "../session/error";
+import { logDebugMessage } from "../../logger";
 
 export function validateAndNormaliseUserInput(
     recipeInstance: Recipe,
@@ -175,7 +177,13 @@ function validateAndNormaliseEmailVerificationConfig(
                           userInfo === undefined ||
                           config?.emailVerificationFeature?.createAndSendCustomEmail === undefined
                       ) {
-                          throw new Error("Unknown User ID provided");
+                          logDebugMessage(
+                              "createAndSendCustomEmail: returning UNAUTHORISED because userInfo is undefined"
+                          );
+                          throw new SessionError({
+                              type: SessionError.UNAUTHORISED,
+                              message: "Unknown User ID provided",
+                          });
                       }
                       return await config.emailVerificationFeature.createAndSendCustomEmail(
                           userInfo,
@@ -195,7 +203,13 @@ function validateAndNormaliseEmailVerificationConfig(
                           userInfo === undefined ||
                           config?.emailVerificationFeature?.getEmailVerificationURL === undefined
                       ) {
-                          throw new Error("Unknown User ID provided");
+                          logDebugMessage(
+                              "getEmailVerificationURL: returning UNAUTHORISED because userInfo is undefined"
+                          );
+                          throw new SessionError({
+                              type: SessionError.UNAUTHORISED,
+                              message: "Unknown User ID provided",
+                          });
                       }
                       return await config.emailVerificationFeature.getEmailVerificationURL(userInfo, userContext);
                   },