+++ lib/ts/recipe/thirdpartyemailpassword/recipe.ts
@@ -39,6 +39,8 @@ import APIImplementation from "./api/implementation";
 import { Querier } from "../../querier";
 import OverrideableBuilder from "supertokens-js-override";
 import EmailDeliveryIngredient from "../../ingredients/emaildelivery";
+import SessionError from "../session/error";
+import { logDebugMessage } from "../../logger";
 
 export default class Recipe extends RecipeModule {
     private static instance: Recipe | undefined = undefined;
@@ -293,7 +295,11 @@ export default class Recipe extends RecipeModule {
     getEmailForUserId = async (userId: string, userContext: any) => {
         let userInfo = await this.recipeInterfaceImpl.getUserById({ userId, userContext });
         if (userInfo === undefined) {
-            throw new Error("Unknown User ID provided");
+            logDebugMessage("getEmailForUserId: returning UNAUTHORISED because userInfo is undefined");
+            throw new SessionError({
+                type: SessionError.UNAUTHORISED,
+                message: "Unknown User ID provided",
+            });
         }
         return userInfo.email;
     };