+++ lib/ts/recipe/thirdpartyemailpassword/utils.ts
@@ -22,6 +22,8 @@ import { normaliseSignUpFormFields } from "../emailpassword/utils";
 import { RecipeInterface, APIInterface } from "./types";
 import BackwardCompatibilityService from "./emaildelivery/services/backwardCompatibility";
 import { RecipeInterface as EPRecipeInterface } from "../emailpassword/types";
+import SessionError from "../session/error";
+import { logDebugMessage } from "../../logger";
 
 export function validateAndNormaliseUserInput(
     recipeInstance: Recipe,
@@ -130,7 +132,13 @@ function validateAndNormaliseEmailVerificationConfig(
                           userInfo === undefined ||
                           config?.emailVerificationFeature?.createAndSendCustomEmail === undefined
                       ) {
-                          throw new Error("Unknown User ID provided");
+                          logDebugMessage(
+                              "createAndSendCustomEmail: returning UNAUTHORISED because userInfo is undefined"
+                          );
+                          throw new SessionError({
+                              type: SessionError.UNAUTHORISED,
+                              message: "Unknown User ID provided",
+                          });
                       }
                       return await config.emailVerificationFeature.createAndSendCustomEmail(
                           userInfo,
@@ -150,7 +158,13 @@ function validateAndNormaliseEmailVerificationConfig(
                           userInfo === undefined ||
                           config?.emailVerificationFeature?.getEmailVerificationURL === undefined
                       ) {
-                          throw new Error("Unknown User ID provided");
+                          logDebugMessage(
+                              "getEmailVerificationURL: returning UNAUTHORISED because userInfo is undefined"
+                          );
+                          throw new SessionError({
+                              type: SessionError.UNAUTHORISED,
+                              message: "Unknown User ID provided",
+                          });
                       }
                       return await config.emailVerificationFeature.getEmailVerificationURL(userInfo, userContext);
                   },